Bad value in handle->request_type for user callback
This is trouble what we have several defines named EVDNS_TYPE_*
or TYPE_* not equal DNS_* and we can't mix them.
May be better remove all of that stuff and prefer one of them.
But now only crutch can be used or revert the code with if expression
if all case of switch for no have_reply value.